+++ src/confcom/azext_confcom/fragment_util.py
@@ -4,6 +4,9 @@
 # --------------------------------------------------------------------------------------------
 
 import yaml
+import copy
+from typing import List
+from knack.log import get_logger
 from azext_confcom import config
 from azext_confcom import oras_proxy
 from azext_confcom.cose_proxy import CoseSignToolProxy
@@ -11,6 +14,8 @@ from azext_confcom.template_util import (
     case_insensitive_dict_get,
     extract_containers_from_text,
 )
+
+logger = get_logger(__name__)
 
 
 # input is the full rego file as a string
@@ -25,31 +30,42 @@ def combine_fragments_with_policy(all_fragments):
     return out_fragments
 
 
-def get_all_fragment_contents(fragment_imports):
+def get_all_fragment_contents(
+    image_names: List[str] = [],
+    fragment_imports: List[dict] = []
+) -> List[str]:
+    # was getting errors with pass by reference so we need to copy it
+    copied_fragment_imports = copy.deepcopy(fragment_imports)
+
     fragment_feeds = [
         case_insensitive_dict_get(fragment, config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_FEED)
-        for fragment in fragment_imports
+        for fragment in copied_fragment_imports
     ]
 
     all_fragments_contents = []
-    cose_proxy = CoseSignToolProxy()
+    # get all the image attached fragments
+    for image in image_names:
+        # TODO: make sure this doesn't error out if the images aren't in a registry. This will probably be in the discover function
+        fragments, feeds = oras_proxy.pull_all_image_attached_fragments(image)
+        for fragment, feed in zip(fragments, feeds):
+            if feed in fragment_feeds:
+                all_fragments_contents.append(fragment)
+            else:
+                logger.warning("Fragment feed %s not in list of feeds to use. Skipping fragment.", feed)
 
-    for fragment in fragment_imports:
+    cose_proxy = CoseSignToolProxy()
+    # get all the local fragments
+    for fragment in copied_fragment_imports:
+        contents = []
         # pull locally if there is a path, otherwise pull from the remote registry
         if (
-            config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_PATH in fragment and
-            fragment[config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_PATH]
+            fragment.get(config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_PATH)
         ):
             contents = [
                 cose_proxy.extract_payload_from_path(
                     fragment[config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_PATH]
                 )
             ]
-        else:
-            feed_name = case_insensitive_dict_get(
-                fragment, config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_FEED
-            )
-            contents = oras_proxy.pull_all_image_attached_fragments(feed_name)
 
         # add the new fragments to the list of all fragments if they're not already there
         # the side effect of adding this way is that if we have a local path to a nested fragment
@@ -68,7 +84,7 @@ def get_all_fragment_contents(fragment_imports):
             # it will end when there are no longer nested fragments to pull
             for new_fragment in fragments:
                 if new_fragment[config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_FEED] not in fragment_feeds:
-                    fragment_imports.append(new_fragment[config.POLICY_FIELD_CONTAINERS_ELEMENTS_REGO_FRAGMENTS_FEED])
+                    copied_fragment_imports.append(new_fragment)
 
             all_fragments_contents.append(content)